VRMShaders.GetTextureParam
