+++ Assets/VRM/Editor/VRM.Editor.asmdef
@@ -7,7 +7,8 @@
         "MeshUtility.Editor",
         "UniUnlit",
         "UniGLTF",
-        "UniGLTF.Editor"
+        "UniGLTF.Editor",
+        "VRMShaders"
     ],
     "optionalUnityReferences": [],
     "includePlatforms": [
