+++ Assets/VRMShaders/Runtime.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fcad5106da772b9478acad653777c2c8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
