+++ Assets/VRMShaders/Runtime/VRMShaders.asmdef
@@ -1,0 +1,3 @@
+﻿{
+	"name": "VRMShaders"
+}
