+++ Assets/VRMShaders/Runtime/VRMShaders.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: da3e51d19d51a544fa14d43fee843098
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
